+++ adev/src/assets/BUILD.bazel
@@ -67,6 +67,7 @@ copy_to_directory(
         "//packages/elements:elements_docs",
         "//packages/forms:forms_docs",
         "//packages/forms/signals:forms_signals_docs",
+        "//packages/forms/signals/compat:forms_signals_compat_docs",
         "//packages/localize:localize_docs",
         "//packages/localize/src/localize:localize_init_docs",
         "//packages/platform-browser:platform-browser_docs",
