+++ goldens/public-api/forms/signals/compat/index.api.md
@@ -1,0 +1,57 @@
+## API Report File for "@angular/forms_signals_compat"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { AbstractControl } from '@angular/forms';
+import * as _angular_forms from '@angular/forms';
+import { ControlValueAccessor } from '@angular/forms';
+import { DestroyableInjector } from '@angular/core';
+import { FormControlStatus } from '@angular/forms';
+import * as i0 from '@angular/core';
+import { Injector } from '@angular/core';
+import { NgControl } from '@angular/forms';
+import { Signal } from '@angular/core';
+import { ValidationErrors } from '@angular/forms';
+import { ValidatorFn } from '@angular/forms';
+import { WritableSignal } from '@angular/core';
+import { ɵCONTROL } from '@angular/core';
+import { ɵControl } from '@angular/core';
+import { ɵFieldState } from '@angular/core';
+import { ɵInteropControl } from '@angular/core';
+
+// @public
+export function compatForm<TModel>(model: WritableSignal<TModel>): FieldTree<TModel>;
+
+// @public
+export function compatForm<TModel>(model: WritableSignal<TModel>, schemaOrOptions: SchemaOrSchemaFn<TModel> | CompatFormOptions): FieldTree<TModel>;
+
+// @public
+export function compatForm<TModel>(model: WritableSignal<TModel>, schema: SchemaOrSchemaFn<TModel>, options: CompatFormOptions): FieldTree<TModel>;
+
+// @public
+export type CompatFormOptions = Omit<FormOptions, 'adapter'>;
+
+// @public
+export class CompatValidationError<T = unknown> implements ValidationError {
+    constructor({ context, kind, control }: {
+        context: T;
+        kind: string;
+        control: AbstractControl;
+    });
+    // (undocumented)
+    readonly context: T;
+    // (undocumented)
+    readonly control: AbstractControl;
+    // (undocumented)
+    readonly field: FieldTree<unknown>;
+    // (undocumented)
+    readonly kind: string;
+    // (undocumented)
+    readonly message?: string;
+}
+
+// (No @packageDocumentation comment for this package)
+
+```
