+++ packages/BUILD.bazel
@@ -108,6 +108,7 @@ filegroup(
         "//packages/elements:elements_docs_extraction",
         "//packages/forms:forms_docs_extraction",
         "//packages/forms/signals:forms_signals_docs_extraction",
+        "//packages/forms/signals/compat:forms_signals_compat_docs_extraction",
         "//packages/localize:localize_docs_extraction",
         "//packages/localize/src/localize:localize_init_docs_extraction",
         "//packages/platform-browser:platform-browser_docs_extraction",
