+++ packages/forms/signals/compat/PACKAGE.md
@@ -1,0 +1,4 @@
+# APIs for Signal Forms interop with Reactive Forms
+
+This package contains APIs to enable interoperability between signal-based forms and reactive forms.
+It allows developers to incrementally migrate their existing forms implementations to signal-based forms.
