+++ packages/forms/signals/test/node/compat/compat_validation_error.spec.ts
@@ -7,10 +7,10 @@
  */
 
 import {FormArray, FormControl, FormGroup, ValidationErrors, ValidatorFn} from '@angular/forms';
+import {CompatValidationError} from '../../../compat/src/api/compat_validation_error';
 import {
-  reactiveErrorsToSignalErrors,
-  CompatValidationError,
   extractNestedReactiveErrors,
+  reactiveErrorsToSignalErrors,
 } from '../../../compat/src/compat_validation_error';
 
 describe('destroy$', () => {
